requirements.txt
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
chardet==3.0.4
geographiclib==1.50
geopy==2.0.0
idna==2.10
jmespath==0.10.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.1
pandas==1.1.2
pgeocode==0.2.1
pip==20.2.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
setuptools==49.6.0.post20200925
six==1.15.0
urllib3==1.25.10
wheel==0.35.1
wincertstore==0.2
